[XMToolbar] Refactor CALayer image setting into extension.
Possibly fix status icon being blurry.
